requirements.txt
# Essential Day 1 Packages (no version pinning to avoid conflicts)
pandas
numpy
requests
matplotlib
seaborn
scikit-learn
python-dotenv
jupyter
tqdm

# Optional but useful for Day 1
scipy
plotly

# AWS Lambda Requirements - Lightweight versions
boto3
psycopg2-binary
numpy

# Database (we'll start with SQLite, add PostgreSQL later if needed)
# psycopg2-binary

# Day 2+ Packages (install when needed)
# kafka-python
# confluent-kafka

# Day 3+ Packages (install when needed) 
# boto3
# botocore

# Day 4+ Packages (install when needed)
# streamlit

# Development tools
# pytest